requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
elasticsearch==7.14.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
